requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\Faculdade\dev\.py\RoleManager\utils\logger.py: 4
colorlog == 6.4.1

# D:\Faculdade\dev\.py\RoleManager\cogs\guildmanager.py: 3,4
# D:\Faculdade\dev\.py\RoleManager\cogs\humanresources.py: 3
# D:\Faculdade\dev\.py\RoleManager\cogs\loadmanager.py: 3
# D:\Faculdade\dev\.py\RoleManager\cogs\rolemanager.py: 3
# D:\Faculdade\dev\.py\RoleManager\cogs\utils.py: 3
# D:\Faculdade\dev\.py\RoleManager\main.py: 8
# D:\Faculdade\dev\.py\RoleManager\utils\converters.py: 2
# D:\Faculdade\dev\.py\RoleManager\utils\embed.py: 1
discord.py == 1.7.3

# D:\Faculdade\dev\.py\RoleManager\cogs\guildmanager.py: 7
# D:\Faculdade\dev\.py\RoleManager\cogs\loadmanager.py: 6
# D:\Faculdade\dev\.py\RoleManager\cogs\rolemanager.py: 14
# D:\Faculdade\dev\.py\RoleManager\main.py: 15
pymongo == 3.12.0

# D:\Faculdade\dev\.py\RoleManager\cogs\guildmanager.py: 18
# D:\Faculdade\dev\.py\RoleManager\cogs\humanresources.py: 17
# D:\Faculdade\dev\.py\RoleManager\cogs\loadmanager.py: 17
# D:\Faculdade\dev\.py\RoleManager\cogs\rolemanager.py: 19
# D:\Faculdade\dev\.py\RoleManager\cogs\utils.py: 17
# D:\Faculdade\dev\.py\RoleManager\main.py: 24
python_dotenv == 0.18.0
